+++ linknamu/src/main/java/com/kakao/linknamu/_core/security/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.kakao.linknamu._core.security;
 
-import com.kakao.linknamu.user.User;
+import com.kakao.linknamu.user.entity.User;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
