+++ linknamu/src/main/java/com/kakao/linknamu/_core/security/JwtProvider.java
@@ -4,7 +4,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.SignatureVerificationException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.kakao.linknamu.user.User;
+import com.kakao.linknamu.user.entity.User;
 import com.auth0.jwt.JWT;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
