+++ linknamu/src/main/java/com/kakao/linknamu/bookmark/BookmarkJPARepository.java
@@ -1,6 +1,6 @@
 package com.kakao.linknamu.bookmark;
 
-import com.kakao.linknamu.user.User;
+import com.kakao.linknamu.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookmarkJPARepository extends JpaRepository<User, Integer> {
