+++ linknamu/src/main/java/com/kakao/linknamu/user/repository/UserJPARepository.java
@@ -1,5 +1,6 @@
-package com.kakao.linknamu.user;
+package com.kakao.linknamu.user.repository;
 
+import com.kakao.linknamu.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
