+++ linknamu/src/main/java/com/kakao/linknamu/user/service/UserService.java
@@ -1,0 +1,13 @@
+package com.kakao.linknamu.user.service;
+
+import com.kakao.linknamu.user.repository.UserJPARepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserService {
+    private final UserJPARepository userJPARepository;
+}
